In signals branch: more work on UDT transport, rpcprotocols and threadpool.
git-svn-id: https://maidsafe-dht.googlecode.com/svn/branches/signals@556 16f77258-3af3-11de-be05-5f7a86268029

+++ src/maidsafe/rpcprotocol/channel.cc
@@ -98,11 +98,11 @@ std::string Controller::method() const {
   return controller_pimpl_->method();
 }
 
-void Controller::set_timeout(const boost::uint64_t &timeout) {
+void Controller::set_timeout(const boost::uint32_t &timeout) {
   controller_pimpl_->set_timeout(timeout);
 }
 
-boost::uint64_t Controller::timeout() const {
+boost::uint32_t Controller::timeout() const {
   return controller_pimpl_->timeout();
 }
